refactor : 엔티티 개선 - 불필요한 연관관계 제거 - RatingId를 내부 클래스로 이동
