+++ src/main/java/app/bottlenote/rating/domain/Rating.java
@@ -1,11 +1,17 @@
 package app.bottlenote.rating.domain;
 
+import static lombok.AccessLevel.PROTECTED;
+
 import app.bottlenote.common.domain.BaseEntity;
 import jakarta.persistence.Column;
+import jakarta.persistence.Embeddable;
 import jakarta.persistence.Embedded;
 import jakarta.persistence.EmbeddedId;
 import jakarta.persistence.Entity;
+import java.io.Serializable;
+import java.util.Objects;
 import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -35,4 +41,38 @@ public class Rating extends BaseEntity {
 		ratingPoint.isValidRating(ratingPoint.getRating());
 		this.ratingPoint = ratingPoint;
 	}
+
+	@Getter
+	@AllArgsConstructor
+	@NoArgsConstructor(access = PROTECTED)
+	@Embeddable
+	public static class RatingId implements Serializable {
+
+		@Column(name = "user_id")
+		private Long userId;
+		@Column(name = "alcohol_id")
+		private Long alcoholId;
+
+		@Override
+		public boolean equals(Object o) {
+			if (this == o) {
+				return true;
+			}
+			if (o == null || getClass() != o.getClass()) {
+				return false;
+			}
+			RatingId ratingId = (RatingId) o;
+			return Objects.equals(userId, ratingId.userId) && Objects.equals(alcoholId, ratingId.alcoholId);
+		}
+
+		@Override
+		public int hashCode() {
+			return Objects.hash(userId, alcoholId);
+		}
+
+		public static RatingId is(Long userId, Long alcoholId) {
+			return new RatingId(userId, alcoholId);
+		}
+	}
+
 }
