+++ src/main/java/app/bottlenote/rating/repository/JpaRatingRepository.java
@@ -1,15 +1,14 @@
 package app.bottlenote.rating.repository;
 
 import app.bottlenote.rating.domain.Rating;
-import app.bottlenote.rating.domain.RatingId;
+import app.bottlenote.rating.domain.Rating.RatingId;
 import app.bottlenote.rating.domain.RatingRepository;
 import app.bottlenote.rating.dto.response.UserRatingResponse;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 public interface JpaRatingRepository extends RatingRepository, JpaRepository<Rating, RatingId>, CustomRatingQueryRepository {
