+++ src/main/java/app/bottlenote/rating/service/RatingCommandService.java
@@ -1,8 +1,10 @@
 package app.bottlenote.rating.service;
 
+import static java.lang.Boolean.FALSE;
+
 import app.bottlenote.alcohols.service.domain.AlcoholFacade;
 import app.bottlenote.rating.domain.Rating;
-import app.bottlenote.rating.domain.RatingId;
+import app.bottlenote.rating.domain.Rating.RatingId;
 import app.bottlenote.rating.domain.RatingPoint;
 import app.bottlenote.rating.domain.RatingRepository;
 import app.bottlenote.rating.dto.payload.RatingRegistryEvent;
@@ -14,13 +16,10 @@ import app.bottlenote.user.exception.UserException;
 import app.bottlenote.user.exception.UserExceptionCode;
 import app.bottlenote.user.service.UserFacade;
 import jakarta.transaction.Transactional;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
-
-import java.util.Objects;
-
-import static java.lang.Boolean.FALSE;
 
 @Slf4j
 @Service
@@ -58,7 +57,7 @@ public class RatingCommandService {
 
 		if (rating == null) {
 			rating = Rating.builder()
-				.id(RatingId.is(userId, alcoholId))
+				.id(new RatingId(userId, alcoholId))
 				.ratingPoint(ratingPoint)
 				.build();
 		} else {
