+++ src/test/java/app/bottlenote/rating/fixture/InMemoryRatingRepository.java
@@ -2,12 +2,11 @@ package app.bottlenote.rating.fixture;
 
 import app.bottlenote.global.service.cursor.PageResponse;
 import app.bottlenote.rating.domain.Rating;
-import app.bottlenote.rating.domain.RatingId;
+import app.bottlenote.rating.domain.Rating.RatingId;
 import app.bottlenote.rating.domain.RatingRepository;
 import app.bottlenote.rating.dto.dsl.RatingListFetchCriteria;
 import app.bottlenote.rating.dto.response.RatingListFetchResponse;
 import app.bottlenote.rating.dto.response.UserRatingResponse;
-
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
