+++ src/test/java/app/bottlenote/rating/fixture/RatingTestFactory.java
@@ -2,7 +2,7 @@ package app.bottlenote.rating.fixture;
 
 import app.bottlenote.alcohols.domain.Alcohol;
 import app.bottlenote.rating.domain.Rating;
-import app.bottlenote.rating.domain.RatingId;
+import app.bottlenote.rating.domain.Rating.RatingId;
 import app.bottlenote.rating.domain.RatingPoint;
 import app.bottlenote.rating.domain.RatingRepository;
 import app.bottlenote.user.domain.User;
